// src/models/case.ts
import mongoose from "mongoose";
const Schema = mongoose.Schema;

interface ICase {
  userID: string;
  warning: string;
  userPhone: string;
  snoozedUntil: Date;
}

//Autogenerated by app. Use to create new case.
const caseSchema = new Schema<ICase>(
  {
    // user ID
    userID: { type: String },
    // type of warning (high, low, OOR, etc.)
    warning: { type: String },
    // user's phone number
    userPhone: { type: String },
    // used in checker.js
    snoozedUntil: { type: Date, default: null },
  },

  { timestamps: true }
);

const Case = mongoose.model("Case", caseSchema);
export default Case;
export { ICase };